Fix rangeOfCode and rangeOfStringLiteral
Before, it'd not honor the startIndex parameter properly in the case a
sectionRange was overlapping it, and a match was before the actual startIndex.
